server/requirements.txt
asgiref==3.8.1
Django==5.2.2
djangorestframework==3.16.0
sqlparse==0.5.3
django-extensions==4.1
requests==2.32.3
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
python-dotenv==1.1.0